+++ lib/riverpod_infinite_scroll_pagination.dart
@@ -5,6 +5,7 @@ export 'src/classes/paginated_data_repository.dart';
 export 'src/interfaces/paginated_notifier.dart';
 export 'src/mixins/keep_alive_paginated_data_mixin.dart';
 export 'src/mixins/paginated_data_mixin.dart';
+export 'src/mixins/paginated_data_mixin_generic.dart';
 export 'src/mixins/paginated_scroll_controller.dart';
 export 'src/models/paginated_response.dart';
 export 'src/models/pagination.dart';