+++ lib/src/models/pagination.g.dart
@@ -9,7 +9,6 @@ part of 'pagination.dart';
 _$PaginationImpl _$$PaginationImplFromJson(Map<String, dynamic> json) =>
     _$PaginationImpl(
       totalNumber: json['total_number'] as int? ?? 0,
-      perPage: json['per_page'] as int? ?? 10,
       currentPage: json['current_page'] as int? ?? 1,
       lastPage: json['last_page'] as int? ?? 1,
     );
@@ -17,7 +16,6 @@ _$PaginationImpl _$$PaginationImplFromJson(Map<String, dynamic> json) =>
 Map<String, dynamic> _$$PaginationImplToJson(_$PaginationImpl instance) =>
     <String, dynamic>{
       'total_number': instance.totalNumber,
-      'per_page': instance.perPage,
       'current_page': instance.currentPage,
       'last_page': instance.lastPage,
     };