+++ CHANGELOG.md
@@ -1,7 +1,12 @@
+## 1.0.2
+
+- Documentation updates
+- Added example for PaginatedDataMixinGeneric (When you need to accept parameters inside build() method)
+
 ## 1.0.1
 
-* Documentation updates and some linter warning fixes
+- Documentation updates and some linter warning fixes
 
 ## 1.0.0
 
-* Initial release.
+- Initial release.